requirements.txt
### APP
Flask==2.0.2
Werkzeug==2.0.1
gunicorn==20.1.0

### ML
tensorflow==2.5.0
numpy==1.19.5
opencv_python==4.5.3.56
Pillow==8.4.0
matplotlib

### DEPLOY
google-cloud-storage==1.42.3